+++ CHANGELOG.md
@@ -11,6 +11,10 @@
 
 - improved message while loading the workspace
 
+### Fixed
+
+- URI protocol handler is now able to switch to the Coder provider even if the last opened provider was something else  
+
 ## 0.3.2 - 2025-06-25
 
 ### Changed